+++ src/PsSpecialHandler.h
@@ -83,8 +83,8 @@ class PsSpecialHandler : public SpecialHandler, protected PSActions
 		void setdash (std::vector<double> &p);
 		void setgray (std::vector<double> &p);
 		void sethsbcolor (std::vector<double> &hsb);
-		void setlinecap (std::vector<double> &p)     {_linecap = (unsigned int) p[0];}
-		void setlinejoin (std::vector<double> &p)    {_linejoin = (unsigned int) p[0];}
+		void setlinecap (std::vector<double> &p)     {_linecap = UInt8(p[0]);}
+		void setlinejoin (std::vector<double> &p)    {_linejoin = UInt8(p[0]);}
 		void setlinewidth (std::vector<double> &p)   {_linewidth = p[0] ? p[0]*1.00375 : 0.5;}
 		void setmatrix (std::vector<double> &p);
 		void setmiterlimit (std::vector<double> &p)  {_miterlimit = p[0]*1.00375;}
@@ -102,11 +102,11 @@ class PsSpecialHandler : public SpecialHandler, protected PSActions
 		XMLElementNode *_xmlnode;   ///< if != 0, created SVG elements are appended to this node
 		GraphicPath<double> _path;
 		DPair _currentpoint;        ///< current PS position
-		float _linewidth;           ///< current linewidth
-		float _miterlimit;          ///< current miter limit
-		float _opacityalpha;        ///< opacity level (0=fully transparent, ..., 1=opaque)
-		unsigned _linecap  : 2;     ///< current line cap (0=butt, 1=round, 2=projecting square)
-		unsigned _linejoin : 2;     ///< current line join (0=miter, 1=round, 2=bevel)
+		double _linewidth;          ///< current linewidth
+		double _miterlimit;         ///< current miter limit
+		double _opacityalpha;       ///< opacity level (0=fully transparent, ..., 1=opaque)
+		UInt8 _linecap  : 2;        ///< current line cap (0=butt, 1=round, 2=projecting square)
+		UInt8 _linejoin : 2;        ///< current line join (0=miter, 1=round, 2=bevel)
 		int _dashoffset;            ///< current dash offset
 		std::vector<int> _dashpattern;
 		ClippingStack _clipStack;
